+++ frontend/js/buypage.js
@@ -1,294 +1,4 @@
-
-let carrouselSalesInfo = [
-    {
-        "itemsPerPage": 5,
-        "title": "Interesses",
-        "items": [
-            { "curQtty": 3, "milestones": [{ "quantity": 0, "price": 50 },{ "quantity": 50, "price": 40 },{ "quantity": 75, "price": 25 }, { "quantity": 100, "price": 15 }], "alt": "item0", "id": "aid1c0", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "aid1c1", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/300" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "aid1c2", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/400" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "aid1c3", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/500" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "aid1c4", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/600" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "aid1c5", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/700" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "aid1c6", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/800" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "aid1c7", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/900" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "aid1c8", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/1000" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "aid1c9", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/1100" }
-
-        ]
-    },
-
-    {
-        "itemsPerPage": 2,
-        "title": "Para você",
-        "items": [
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item0", "id": "aid2c0", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "aid2c1", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 1, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "aid2c2", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 9, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "aid2c3", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 7, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "aid2c4", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "aid2c5", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 3, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "aid2c6", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "aid2c7", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "aid2c8", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 0, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "aid2c9", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" }
-
-        ]
-    },
-    {
-        "itemsPerPage": 6,
-        "title": "Próximos de esgotar",
-        "items": [
-            { "curQtty": 4, "milestones": [{ "quantity": 30, "price": 25 }, { "quantity": 500, "price": 15 }], "alt": "item0", "id": "aid3c0", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "aid3c1", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 1, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "aid3c2", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 9, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "aid3c3", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 7, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "aid3c4", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "aid3c5", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 3, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "aid3c6", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "aid3c7", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "aid3c8", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 0, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "aid3c9", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" }
-
-        ]
-    },
-    {
-        "itemsPerPage": 1,
-        "title": "Em promoção",
-        "items": [
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item0", "id": "aid4c0", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "aid4c1", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 1, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "aid4c2", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 9, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "aid4c3", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 7, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "aid4c4", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "aid4c5", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 3, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "aid4c6", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "aid4c7", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "aid4c8", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 0, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "aid4c9", "link": "../pages/buypage.html", "imageSrc": "https://picsum.photos/200" }
-
-        ]
-    },
-];
-function getItemAttrs(item) {
-    let minPriceMile = { price: item.milestones[0].price + 1 };
-    let curPrice;
-    let curQtty = item.curQtty;
-    let nearestMile = { "quantity": -1 };
-    let biggestMile = { "quantity": -1 };
-    for (milestone of item.milestones) {
-        if (milestone.quantity > biggestMile.quantity)
-        biggestMile.quantity = milestone.quantity;
-
-        if (curQtty > milestone.quantity && milestone.quantity > nearestMile.quantity)
-            nearestMile = milestone;
-            
-        if (milestone.price < minPriceMile.price) minPriceMile = milestone;
-    }
-    
-    if (biggestMile.quantity === -1) {
-        console.error('Invalid product: no valid milestones. ID = ', item.id);
-    }
-
-    if (nearestMile.quantity === -1) curPrice = '-';
-    else curPrice = nearestMile.price;
-    
-    return { curQtty, maxQuantity: biggestMile.quantity, curPrice, minPrice: minPriceMile.price };
-}
-/*
-
-function generateItem(item) {
-    let itemAttrs = getItemAttrs(item);
-    return `
-        <div class="col mx-auto" style="min-width:00px; max-width:200px">
-            <div class="card">
-                <a href=${item.link}>
-                    <img src="${item.imageSrc}" class="d-block img-thumbnail mx-auto" alt="${item.alt}" />
-                </a>
-                <div class="card-body px-1 pt-2">
-                    <div class="row g-0 p-0">
-                        <div class="col-10" style="text-transform:capitalize;"> ${item.alt} </div>
-                    </div>
-                    <div class="row flex-direction-row mt-2 px-1">
-                        <div class="col-6">
-                            R$${itemAttrs.curPrice}
-                        </div>
-                        <div class="col-6 text-end">
-                            R$${itemAttrs.minPrice}
-                        </div>
-                    </div>
-                    <div class="progress">
-                        <div class="progress-bar text-center bg-warning" role="progressbar" style="width: ${100 * itemAttrs.curQtty / itemAttrs.maxQuantity}%" aria-valuenow="10" aria-valuemin="0" aria-valuemax="100">
-                        <small class="justify-content-center d-flex position-absolute w-100 text-dark fw-bold">${itemAttrs.curQtty}/${itemAttrs.maxQuantity}</small>
-                        
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    `
-}
-*/
-
-
-
-
-
-function workOnMilestones(milestoneCur){
-    let milestoneIsActive = milestoneCur.hasClass('active');
-
-    let milestoneTargets = $('#milestone-list .milestone-item')
-
-    if (!milestoneIsActive) {
-        milestoneTargets.addClass('d-none');
-        milestoneCur.removeClass('d-none');
-
-        milestoneTargets.find('.card-body').addClass('d-none');
-        milestoneCur.find('.card-body').removeClass('d-none');
-
-        milestoneCur.addClass('active');
-        milestoneCur.addClass('h-100');
-    } else {
-        milestoneTargets.removeClass('d-none');
-        milestoneTargets.find('.card-body').addClass('d-none');
-        
-        milestoneCur.removeClass('active');
-        milestoneCur.removeClass('h-100');
-    }
-}
-
-function workOnMilestoneSpots(milestoneSpotCur){
-    let milestoneSpotIsActive = milestoneSpotCur.hasClass('active');
-
-    let milestoneSpotTargets = $('.milestone-progress-spot')
-
-    if (!milestoneSpotIsActive) {
-        milestoneSpotTargets.parent().parent().addClass('d-none');
-        milestoneSpotCur.removeClass('d-none');
-
-        milestoneSpotCur.addClass('active');
-    } else {
-        milestoneSpotTargets.parent().parent().removeClass('d-none');            
-        milestoneSpotCur.removeClass('active');
-    }
-}
-
-
-
-
-
-
-
 $(document).ready(function () {
-    /*
-    async function populateCarrousel(carrouselType, location, idName, generateItem, classes) {
-        classes = classes || '';
-
-        let itemsPerPage = Math.min(carrouselType.itemsPerPage, carrouselType.items.length);
-        if (itemsPerPage <= 0) return;
-
-        let newHtml = `<div class="row layer ${classes} carousel-container mt-2" id="carousel-${idName}-row">
-                        <h4 class="${classes || 'mt-5'} w-100 text-center ${classes}  ">${carrouselType.title}</h4>
-                        <div class="col-9 mx-auto w-100">
-                            <div class="carousel slide" id="carousel-${idName}" data-bs-ride="carousel">
-                                <div class="carousel-inner w-75 mx-auto">`;
-
-
-        let pageCount = Math.ceil(carrouselType.items.length / itemsPerPage);
-        let lastPageItemsToAdd = (itemsPerPage - carrouselType.items.length % itemsPerPage) % itemsPerPage;
-        for (let i = 0; i < pageCount; i++) {
-            newHtml += `<div class="${'carousel-item row-car ' + (i === 0 ? 'active' : '')} ">
-                            <div class="row g-1">`
-            let k = i * itemsPerPage;
-
-            if (i === pageCount - 1) k -= lastPageItemsToAdd;
-
-            for (let j = 0; j < itemsPerPage; j++) {
-                let item = carrouselType.items[k + j];
-                newHtml += generateItem(item);
-            }
-            newHtml += `
-                            </div>
-                        </div>
-            `;
-        }
-        newHtml += "</div>";
-        newHtml += `
-                    <button class="carousel-control-prev" type="button" data-bs-target="#carousel-${idName}" data-bs-slide="prev">
-                        <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-                        <span class="visually-hidden">Previous</span>
-                    </button>
-                    <button class="carousel-control-next" type="button" data-bs-target="#carousel-${idName}" data-bs-slide="next">
-                        <span class="carousel-control-next-icon" aria-hidden="true"></span>
-                        <span class="visually-hidden">Next</span>
-                    </button>
-        `
-
-        newHtml += `
-                        </div>
-                    </div>
-                </div>
-            `;
-
-        let currHtml = $(location).html();
-        $(location).html(currHtml + newHtml);
-
-    }
-    */
-
-    function fixProgressBar(product){
-        milestoneAttrs = getItemAttrs(product);
-        $('.progress-bar').attr("style",`width: ${100*(product.curQtty/milestoneAttrs.maxQuantity)}%`);
-        $('.progress-bar small').html(`${product.curQtty}/${milestoneAttrs.maxQuantity}`);
-    }
-
-    function populateMilestones(product) {
-        let newHtml = ''
-
-        let milestoneInfo = product.milestones;
-
-        let i =  1
-
-        fixProgressBar(product);
-        for (let milestone of milestoneInfo) {
-            newHtml += `
-            <div class="milestone-item row card mx-auto m-2 p-1 w-100" data-milestone="${product.alt}-${milestone.quantity}">
-                <div class="card-header bg-transparent">
-                    Meta: ${product.curQtty}/${milestone.quantity} 
-                </div>
-                <div class="card-body d-none">
-                    <h5 class="card-title">R$${milestone.price}</h5>
-                    <p class="card-text d-none">With supporting text below as a natural lead-in to additional content.</p>
-                </div>
-            </div>
-            `
-        }
-
-
-        let currHtml = $('#milestone-list').html();
-        $('#milestone-list').html(currHtml + newHtml);
-    }
-
-    function populateMilestoneSpots(product){
-        newHtml = '';
-        let mileSpotsAttrs;
-        let milestones = product.milestones;
-        for(let milestone of milestones){
-            mileSpotsAttrs = getItemAttrs(product);
-            newHtml += `
-                <div class="w-100 noclick d-flex position-absolute start-0">
-                    <span class="invisible noclick" style="width:${100 * (1/6 +  (2 * milestone.quantity) / ( 3 * mileSpotsAttrs.maxQuantity ))}%;"></span>
-                    <a data-milestone="${product.alt}-${milestone.quantity}" href="">
-                        <span class="milestone-progress-spot fa noclick fa-caret-up"></span>
-                    </a>
-                </div> 
-            `;
-        }
-        let currHtml = $('#progress-spots').html();
-        $('#progress-spots').html(currHtml + newHtml);
-    }
-
-
     populateMilestones(carrouselSalesInfo[0].items[0]);
     populateMilestoneSpots(carrouselSalesInfo[0].items[0]);
 
@@ -309,53 +19,4 @@ $(document).ready(function () {
         workOnMilestoneSpots($(this).parent());
         workOnMilestones($(`div[data-milestone=${$(this).attr("data-milestone")}]`));
     });
-
-
-
-    /*
-    $('.milestone-item .card-header').click(function (e) {
-        e.preventDefault();
-        console.log('asas')
-
-        //Hides every other milestone in the list
-        let self = $(this).parent();
-
-        let isActive = self.hasClass('active');
-
-        let targets = $('#milestone-list .milestone-item')
-
-        if (!isActive) {
-            targets.addClass('d-none');
-            self.removeClass('d-none');
-
-                
-            targets.find('.card-body').addClass('d-none');
-            self.find('.card-body').removeClass('d-none');
-
-            self.addClass('active');
-            self.addClass('h-100');
-        } else {
-            targets.removeClass('d-none');
-            targets.find('.card-body').addClass('d-none');
-            
-            self.removeClass('active');
-            self.removeClass('h-100');
-        }
-
-    });*/
-
-
-    /*
-
-    $('#categories').mouseleave(function (e) {
-        e.preventDefault();
-
-        let layers = $(this).find(".layer[data-layer]:not([data-layer='1']")
-        layers.addClass('d-none');
-        layers.find('.category').addClass('d-none');
-        $('#carousel-layer-row').addClass('d-none');
-
-
-    });
-    */
 });
