+++ web-group-buying-store/src/components/ProductCard.tsx
@@ -4,12 +4,12 @@ import { Link } from 'react-router-dom';
 import 'bootstrap/dist/css/bootstrap.min.css'
 import 'bootstrap/dist/js/bootstrap.min.js'
 
-type Milestone = {
+export type MilestoneProps = {
     quantity: number,
     price: number,
 };
 
-type UserProps = {
+export type UserProps = {
     name: string,
     nick: string,
     email: string,
@@ -31,7 +31,7 @@ export type ProductProps = {
     description?: string,
     imageURL: string,
     category: string,
-    milestones: Milestone[];
+    milestones: MilestoneProps[];
     currentQuantity: number,
     comments?: ProductCommentProps[],
 };
